Move bank end to end tests
