+++ tests/e2e/bank/client/cli_helpers.go
@@ -1,4 +1,4 @@
-package testutil
+package client
 
 import (
 	"fmt"
