+++ tests/e2e/bank/client/cli_test.go
@@ -1,14 +1,13 @@
 //go:build e2e
 // +build e2e
 
-package testutil
+package client
 
 import (
 	"testing"
 
 	"github.com/cosmos/cosmos-sdk/simapp"
 	"github.com/cosmos/cosmos-sdk/testutil/network"
-	"github.com/cosmos/cosmos-sdk/x/bank/client/testutil"
 
 	"github.com/stretchr/testify/suite"
 )
@@ -16,5 +15,5 @@ import (
 func TestIntegrationTestSuite(t *testing.T) {
 	cfg := network.DefaultConfig(simapp.NewTestNetworkFixture)
 	cfg.NumValidators = 1
-	suite.Run(t, testutil.NewIntegrationTestSuite(cfg))
+	suite.Run(t, NewEndToEndTestSuite(cfg))
 }
