+++ tests/e2e/bank/client/suite.go
@@ -1,14 +1,15 @@
-package testutil
+package client
 
 import (
 	"fmt"
 	"io/ioutil"
 	"os"
 
-	"cosmossdk.io/math"
 	"github.com/gogo/protobuf/proto"
 	"github.com/stretchr/testify/suite"
 	tmcli "github.com/tendermint/tendermint/libs/cli"
+
+	"cosmossdk.io/math"
 
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
@@ -20,18 +21,18 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/bank/types"
 )
 
-type IntegrationTestSuite struct {
+type EndToEndTestSuite struct {
 	suite.Suite
 
 	cfg     network.Config
 	network *network.Network
 }
 
-func NewIntegrationTestSuite(cfg network.Config) *IntegrationTestSuite {
-	return &IntegrationTestSuite{cfg: cfg}
+func NewEndToEndTestSuite(cfg network.Config) *EndToEndTestSuite {
+	return &EndToEndTestSuite{cfg: cfg}
 }
 
-func (s *IntegrationTestSuite) SetupSuite() {
+func (s *EndToEndTestSuite) SetupSuite() {
 	s.T().Log("setting up integration test suite")
 
 	genesisState := s.cfg.GenesisState
@@ -90,12 +91,12 @@ func (s *IntegrationTestSuite) SetupSuite() {
 	s.Require().NoError(err)
 }
 
-func (s *IntegrationTestSuite) TearDownSuite() {
+func (s *EndToEndTestSuite) TearDownSuite() {
 	s.T().Log("tearing down integration test suite")
 	s.network.Cleanup()
 }
 
-func (s *IntegrationTestSuite) TestGetBalancesCmd() {
+func (s *EndToEndTestSuite) TestGetBalancesCmd() {
 	val := s.network.Validators[0]
 
 	testCases := []struct {
@@ -166,7 +167,7 @@ func (s *IntegrationTestSuite) TestGetBalancesCmd() {
 	}
 }
 
-func (s *IntegrationTestSuite) TestGetCmdQueryTotalSupply() {
+func (s *EndToEndTestSuite) TestGetCmdQueryTotalSupply() {
 	val := s.network.Validators[0]
 
 	testCases := []struct {
@@ -238,7 +239,7 @@ func (s *IntegrationTestSuite) TestGetCmdQueryTotalSupply() {
 	}
 }
 
-func (s *IntegrationTestSuite) TestGetCmdQueryDenomsMetadata() {
+func (s *EndToEndTestSuite) TestGetCmdQueryDenomsMetadata() {
 	val := s.network.Validators[0]
 
 	testCases := []struct {
@@ -365,7 +366,7 @@ func (s *IntegrationTestSuite) TestGetCmdQueryDenomsMetadata() {
 	}
 }
 
-func (s *IntegrationTestSuite) TestNewSendTxCmdGenOnly() {
+func (s *EndToEndTestSuite) TestNewSendTxCmdGenOnly() {
 	val := s.network.Validators[0]
 
 	clientCtx := val.ClientCtx
@@ -390,7 +391,7 @@ func (s *IntegrationTestSuite) TestNewSendTxCmdGenOnly() {
 	s.Require().Equal([]sdk.Msg{types.NewMsgSend(from, to, amount)}, tx.GetMsgs())
 }
 
-func (s *IntegrationTestSuite) TestNewSendTxCmdDryRun() {
+func (s *EndToEndTestSuite) TestNewSendTxCmdDryRun() {
 	val := s.network.Validators[0]
 
 	clientCtx := val.ClientCtx
@@ -422,7 +423,7 @@ func (s *IntegrationTestSuite) TestNewSendTxCmdDryRun() {
 	s.Require().Regexp("gas estimate: [0-9]+", string(out))
 }
 
-func (s *IntegrationTestSuite) TestNewSendTxCmd() {
+func (s *EndToEndTestSuite) TestNewSendTxCmd() {
 	val := s.network.Validators[0]
 
 	testCases := []struct {
@@ -524,7 +525,7 @@ func (s *IntegrationTestSuite) TestNewSendTxCmd() {
 	}
 }
 
-func (s *IntegrationTestSuite) TestNewMultiSendTxCmd() {
+func (s *EndToEndTestSuite) TestNewMultiSendTxCmd() {
 	val := s.network.Validators[0]
 	testAddr := sdk.AccAddress("cosmos139f7kncmglres2nf3h4hc4tade85ekfr8sulz5")
